@@ -9,11 +9,7 @@ import { NavComponent } from './surface/nav/nav.component';
 import { HeaderComponent } from './surface/header/header.component';
 import { FooterComponent } from './surface/footer/footer.component';
 import { NgbDateParserFormatter, NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import {
-  HTTP_INTERCEPTORS,
-  HttpClient,
-  HttpClientModule,
-} from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import {
   AuthInterceptor,
   ResponseInterceptor,
@@ -60,54 +56,48 @@ export function HttpLoaderFactory(httpClient: HttpClient) {
   return new TranslateHttpLoader(httpClient, './assets/i18n/', '.json');
 }
 
-@NgModule({
-  declarations: [
-    ErrorComponent,
-    LoginComponent,
-    AppComponent,
-    HomeComponent,
-    MainComponent,
-    NavComponent,
-    HeaderComponent,
-    FooterComponent,
-    DashboardHomeComponent,
-    SearchComponent,
-  ],
-  imports: [
-    NgbModule,
-    FormsModule,
-    BrowserModule,
-    AppRoutingModule,
-    ToastModule,
-    IconsModule,
-    SpinnerModule,
-    FormsModule,
-    HttpClientModule,
-    AddressModule,
-    CommonModule,
-    ModalModule,
-    ScheduleModule,
-    AbsenceGanttModule,
-    TranslateModule.forRoot({
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient],
-      },
-    }),
-  ],
-  providers: [
-    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
-    { provide: HTTP_INTERCEPTORS, useClass: ResponseInterceptor, multi: true },
-    { provide: AppErrorHandler, useClass: AppErrorHandler },
-    { provide: CanDeactivateGuard, useClass: CanDeactivateGuard },
-    { provide: NgbDateParserFormatter, useClass: NgbDateCustomParserFormatter },
-    CurrencyPipe,
-    Title,
-    TranslatePipe,
-    TranslateStringConstantsService,
-    { provide: LOCALE_ID, useValue: 'en-US' },
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        ErrorComponent,
+        LoginComponent,
+        AppComponent,
+        HomeComponent,
+        MainComponent,
+        NavComponent,
+        HeaderComponent,
+        FooterComponent,
+        DashboardHomeComponent,
+        SearchComponent,
+    ],
+    bootstrap: [AppComponent], imports: [NgbModule,
+        FormsModule,
+        BrowserModule,
+        AppRoutingModule,
+        ToastModule,
+        IconsModule,
+        SpinnerModule,
+        FormsModule,
+        AddressModule,
+        CommonModule,
+        ModalModule,
+        ScheduleModule,
+        AbsenceGanttModule,
+        TranslateModule.forRoot({
+            loader: {
+                provide: TranslateLoader,
+                useFactory: HttpLoaderFactory,
+                deps: [HttpClient],
+            },
+        })], providers: [
+        { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+        { provide: HTTP_INTERCEPTORS, useClass: ResponseInterceptor, multi: true },
+        { provide: AppErrorHandler, useClass: AppErrorHandler },
+        { provide: CanDeactivateGuard, useClass: CanDeactivateGuard },
+        { provide: NgbDateParserFormatter, useClass: NgbDateCustomParserFormatter },
+        CurrencyPipe,
+        Title,
+        TranslatePipe,
+        TranslateStringConstantsService,
+        { provide: LOCALE_ID, useValue: 'en-US' },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule {}
